fix: revert to scrypt-js since noble is incompatible
