+++ packages/crypto/src/crypto-browser.ts
@@ -4,10 +4,11 @@ import { hmac } from "@noble/hashes/hmac";
 import { sha512 } from "@noble/hashes/sha512";
 import { sha256 } from "@noble/hashes/sha256";
 import { ripemd160 } from "@noble/hashes/ripemd160";
-import { scrypt } from "@noble/hashes/scrypt";
+// the noble/hashes is not compatible with the https://ethereum.org/en/developers/docs/data-structures-and-encoding/web3-secret-storage/#scrypt
+// import { scrypt } from "@noble/hashes/scrypt";
 import { randomBytes as _randomBytes } from "@noble/hashes/utils";
 import { pbkdf2 as _pbkdf2 } from "@noble/hashes/pbkdf2";
-import type { Hash, HashAlgo, ScryptOptions } from "./types";
+import type { Hash, HashAlgo } from "./types";
 // TODO may replace with the crypto.subtle API
 import {
   createCipheriv as _createCipheriv,
@@ -28,20 +29,6 @@ export function createHash(algorithm: HashAlgo): Hash {
 
 export function createHmac(algorithm: HashAlgo, key: Uint8Array): Hash {
   return hmac.create(_createHash(algorithm), key);
-}
-
-export function scryptSync(
-  password: Uint8Array,
-  salt: Uint8Array,
-  dkLen: number,
-  options: ScryptOptions
-): Uint8Array {
-  return scrypt(password, salt, {
-    dkLen,
-    r: options.r,
-    p: options.p,
-    N: options.N,
-  });
 }
 
 export function randomBytes(bytesLen: number): Uint8Array {
@@ -76,3 +63,5 @@ export function pbkdf2Sync(
     c: iterations,
   });
 }
+
+export { scryptSync, scrypt } from "./scrypt";
